+++ src/worldObject/buildings/Guild.java
@@ -8,7 +8,7 @@ public class Guild extends BaseObject {
     Image i = new Image("/object/House(Cropped).png");
     // x = 582 y = 3 x2 = 913  y2 = 334
     WritableImage image = new WritableImage(i.getPixelReader(),582,3,331,332);
-    public Guild(int WorldX,int WorldY,double scaleX,double scaleY,double solidX,double solidY,double solidW,double solidH){
+    public Guild(double WorldX,double WorldY,double scaleX,double scaleY,double solidX,double solidY,double solidW,double solidH){
         super(WorldX,WorldY,scaleX,scaleY,solidX,solidY,solidW,solidH);
         setName("Guild");
         setCollision(true);
