+++ src/worldObject/buildings/LongHouse.java
@@ -8,7 +8,7 @@ public class LongHouse extends BaseObject {
     Image i = new Image("/object/House(Cropped).png");
     // x = 11 y = 239 x2 = 290  y2 = 433
     WritableImage image = new WritableImage(i.getPixelReader(),11,239,280,195);
-    public LongHouse(int WorldX,int WorldY,double scaleX,double scaleY,double solidX,double solidY,double solidW,double solidH){
+    public LongHouse(double WorldX,double WorldY,double scaleX,double scaleY,double solidX,double solidY,double solidW,double solidH){
         super(WorldX,WorldY,scaleX,scaleY,solidX,solidY,solidW,solidH);
         setName("Long House");
         setCollision(true);
