+++ src/worldObject/forest/ForestTreeB.java
@@ -8,7 +8,7 @@ public class ForestTreeB extends BaseObject {
     Image i = new Image("/object/tree5.png");
     // x = 75, y = 420, x2 = 322, y2 = 731
     WritableImage image = new WritableImage(i.getPixelReader(),75,420,248,312);
-    public ForestTreeB(int WorldX, int WorldY, double scaleX, double scaleY, double solidX, double solidY, double solidW, double solidH){
+    public ForestTreeB(double WorldX, double WorldY, double scaleX, double scaleY, double solidX, double solidY, double solidW, double solidH){
         super(WorldX,WorldY,scaleX,scaleY,solidX,solidY,solidW,solidH); // 4, 4, 1.25, 3, 1.75, 0.5
         setName("Normal tree");
         setCollision(true);
