+++ src/worldObject/forest/ForestTreeS.java
@@ -8,7 +8,7 @@ public class ForestTreeS extends BaseObject {
     Image i = new Image("/object/tree5.png");
     // x = 647, y = 500, x2 = 801, y2 = 671
     WritableImage image = new WritableImage(i.getPixelReader(),647,500,155,172);
-    public ForestTreeS(int WorldX, int WorldY, double scaleX, double scaleY, double solidX, double solidY, double solidW, double solidH){
+    public ForestTreeS(double WorldX, double WorldY, double scaleX, double scaleY, double solidX, double solidY, double solidW, double solidH){
         super(WorldX,WorldY,scaleX,scaleY,solidX,solidY,solidW,solidH); // 2, 2, 0.5, 1.5, 1, 0.25
         setName("Normal tree");
         setCollision(true);
